Move request specs for Review to correct file
