+++ spec/models/review_spec.rb
@@ -2,7 +2,7 @@
 
 require "rails_helper"
 
-# Model test here, request spec follows below
+# Model test here, requests spec follows below
 RSpec.describe Review, type: :model do
   it "can add a review" do
     review = Fabricate(:review)
@@ -68,114 +68,5 @@ RSpec.describe Review, type: :model do
     end.to raise_error(ActiveRecord::RecordInvalid)
 
     expect(review.facility_reviews.count).to eq(1)
-  end
-end
-
-RSpec.describe Review, type: :request do
-  let(:user) { Fabricate(:user) }
-  let(:review) { Fabricate(:review, user: user) }
-  let(:space) { review.space }
-  let(:facility) { Fabricate(:facility) }
-  let(:facility_review) do
-    Fabricate(
-      :facility_review,
-      space: space,
-      user: user,
-      review: review,
-      facility: facility,
-      experience: :was_not_allowed
-    )
-  end
-
-  before do
-    sign_in user
-    facility_review.reload
-    space.aggregate_facility_reviews
-  end
-
-  it "can load the new review paths" do
-    get new_review_path(space_id: space.id)
-    expect(response).to have_http_status(:success)
-
-    get new_review_with_type_of_contact_path(
-      space_id: space.id,
-      type_of_contact: :not_allowed_to_use
-    )
-    expect(response).to have_http_status(:success)
-
-    get new_review_with_type_of_contact_path(
-      space_id: space.id,
-      type_of_contact: :only_contacted
-    )
-    expect(response).to have_http_status(:success)
-
-    get new_review_with_type_of_contact_path(
-      space_id: space.id,
-      type_of_contact: :been_there
-    )
-    expect(response).to have_http_status(:success)
-  end
-
-  it "can create a new review" do
-    expect(space.reviews.count).to eq(1)
-    post reviews_path, params: {
-      review: {
-        title: "They are all mean to me",
-        type_of_contact: :not_allowed_to_use,
-        space_id: space.id
-      }
-    }
-    follow_redirect!
-    expect(response).to have_http_status(:success)
-    expect(space.reviews.count).to eq(2)
-  end
-
-  it "can load the edit path" do
-    get edit_review_path(review)
-    expect(response).to have_http_status(:success)
-  end
-
-  it "has loaded the facility review, and set it to unlikely" do
-    expect(review.facility_reviews.count).to eq(1)
-    expect(space.reviews_for_facility(facility)).to eq("unlikely")
-  end
-
-  it "will update the facility review if the Review is updated" do
-    patch review_path(review), params: {
-      review: {
-        facility_reviews_attributes: {
-          "#{facility.id}": {
-            id: facility_review.id,
-            facility_id: facility.id,
-            experience: "was_allowed"
-          }
-        }
-      }
-    }
-    follow_redirect!
-    expect(response).to have_http_status(:success)
-    space.aggregate_facility_reviews
-    expect(review.facility_reviews.count).to eq(1)
-    expect(space.reviews_for_facility(facility.id)).to eq("likely")
-  end
-
-  it "will delete the facility review if the Review is updated, and it is set to unknown" do
-    expect(review.facility_reviews.count).to eq(1)
-    patch review_path(review), params: {
-      review: {
-        facility_reviews_attributes: {
-          "#{facility.id}": {
-            id: facility_review.id,
-            facility_id: facility.id,
-            experience: "unknown"
-          }
-        }
-      }
-    }
-    follow_redirect!
-    expect(response).to have_http_status(:success)
-    space.aggregate_facility_reviews
-    expect(review.facility_reviews.count).to eq(0)
-    expect(space.reviews_for_facility(facility.id)).to eq("unknown")
   end
 end
